RM-59: Addressed code review comments - moved RemotePatientMatchResource and DefaultRemotePatientMatchResource to a separate rest module.
