+++ core/client/src/main/resources/META-INF/components.txt
@@ -1,4 +1,3 @@
 org.phenotips.remote.client.internal.DefaultRemoteMatchingService
-org.phenotips.remote.client.internal.DefaultRemotePatientMatchResource
 org.phenotips.remote.client.script.RemoteMatchingScriptService
 org.phenotips.remote.client.internal.RemoteMatchFinder
