+++ core/pom.xml
@@ -35,5 +35,6 @@
     <module>common</module>
     <module>server</module>
     <module>client</module>
+    <module>rest</module>
   </modules>
 </project>
