+++ core/rest/src/main/java/org/phenotips/remote/rest/RemotePatientMatchResource.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see http://www.gnu.org/licenses/
  */
-package org.phenotips.remote.client;
+package org.phenotips.remote.rest;
 
 import org.phenotips.data.rest.PatientResource;
 import org.phenotips.rest.ParentResource;
