+++ core/rest/src/main/resources/META-INF/components.txt
@@ -1,0 +1,1 @@
+org.phenotips.remote.rest.internal.DefaultRemotePatientMatchResource
